requirements.txt
streamlit>=1.31
pandas>=2.0
gspread    
google-auth